// src/common/enums/warning.enums.ts
export enum WarningCode {
  SOIL_NOT_RECOMMENDED = 'SOIL_NOT_RECOMMENDED',
  PH_OUT_OF_RANGE = 'PH_OUT_OF_RANGE',
  DEPTH_TOO_SMALL = 'DEPTH_TOO_SMALL',
  NPK_TOO_LOW = 'NPK_TOO_LOW',
  ROTATION_RISK = 'ROTATION_RISK',
  WATER_RETENTION_MISMATCH = 'WATER_RETENTION_MISMATCH',
  DRAINAGE_MISMATCH = 'DRAINAGE_MISMATCH',
  FAMILY_REPETITION = 'FAMILY_REPETITION',
  HARVEST_WINDOW_MISSED = 'HARVEST_WINDOW_MISSED',
  SUBOPTIMAL_SOWING_TIME = 'SUBOPTIMAL_SOWING_TIME',
  EXPERIMENTAL_SETUP = 'EXPERIMENTAL_SETUP',
  FROST_RISK_NEXT_7_DAYS = 'FROST_RISK_NEXT_7_DAYS',
  HARD_FROST_RISK_NEXT_7_DAYS = 'HARD_FROST_RISK_NEXT_7_DAYS',
  DROUGHT_RISK_NEXT_7_DAYS = 'DROUGHT_RISK_NEXT_7_DAYS',
  HEAVY_RAIN_RISK_NEXT_48H = 'HEAVY_RAIN_RISK_NEXT_48H',
  WIND_DAMAGE_RISK_NEXT_48H = 'WIND_DAMAGE_RISK_NEXT_48H',
  FUNGAL_DISEASE_PRESSURE_HIGH = 'FUNGAL_DISEASE_PRESSURE_HIGH',
  OVERWATERING_RISK = 'OVERWATERING_RISK',
  GERMINATION_TOO_COLD = 'GERMINATION_TOO_COLD',
}

export enum WarningScope {
  USER = 'USER',
  BED = 'BED',
  PLANTING = 'PLANTING',
}

export enum WarningSeverity {
  INFO = 'INFO',
  WARNING = 'WARNING',
  CRITICAL = 'CRITICAL',
}